Move the hook into a hooks folder in mgt-reactr
