+++ packages/mgt-react/src/hooks/useIsSignedIn.ts
@@ -1,0 +1,31 @@
+/**
+ * -------------------------------------------------------------------------------------------
+ * Copyright (c) Microsoft Corporation.  All Rights Reserved.  Licensed under the MIT License.
+ * See License in the project root for license information.
+ * -------------------------------------------------------------------------------------------
+ */
+
+import { signedInState, Providers } from '@microsoft/mgt-element';
+import { useState, useEffect } from 'react';
+
+/**
+ * Hook to check if a user is signed in.
+ *
+ * @returns true if a user is signed on, otherwise false.
+ */
+export const useIsSignedIn = (): [boolean] => {
+  const [isSignedIn, setIsSignedIn] = useState(false);
+  useEffect(() => {
+    const updateState = () => {
+      setIsSignedIn(signedInState());
+    };
+
+    Providers.onProviderUpdated(updateState);
+    updateState();
+
+    return () => {
+      Providers.removeProviderUpdatedListener(updateState);
+    };
+  }, [setIsSignedIn]);
+  return [isSignedIn];
+};
